services/face_rec/requirements.txt
fastapi
uvicorn[standard]
redis
pydantic
python-multipart # For FastAPI File Uploads
Pillow # For basic image validation/metadata
async-timeout
